Getting the status of detached processes
With this commit, it is possible to get the status of a service even if
the process wasn't started from emacs.

This can be done using the `live-p` property which is a lambda function
and takes the `service` as the argument.

Let's say the service name is 'hakyll', then the service can be defined
with the `live-p` function as:

```
(prodigy-define-service
  :name "hakyll"
  :command "docker-compose"
  :args '("up")
  :stop-signal 'sigkill
  :kill-process-buffer-on-stop nil
  :live-p (lambda(&rest service)
            (-when-let (name (plist-get service :name))
              (eq 0 (shell-command (concat "docker ps | grep " name) ))
              )))
```